requirements.txt
GitPython==3.1.27
fire~=0.4.0
